requirements.txt
# Core dependencies
python-dotenv
requests
supabase
resend

# LLM related
litellm
tiktoken

# Web scraping and parsing
beautifulsoup4
playwright

# PDF and document processing
PyMuPDF  # for fitz
Pillow   # for PIL/Image

# Git and visualization
GitPython
treelib
boto3

# API clients
arxiv    # Changed back to 'arxiv'
praw    # for Reddit

# Optional but recommended
pytest   # for testing